ensure outputs are native python types

+++ src/distinct.py
@@ -5,14 +5,14 @@ import numpy as np
 def count_distinct(column: pd.Series) -> dict:
     distinct_values = column.nunique()
 
-    return {"distinct": {"count": distinct_values,
-                              "proportion": distinct_values / len(column)}}
+    return {"distinct": {"count": int(distinct_values),
+                              "proportion": float(distinct_values / len(column))}}
 
 
 def count_zero_values(column: pd.Series) -> dict:
     num_zero = np.sum(np.array(column == 0))
 
-    return {"zeros": {"count": num_zero,
-                           "proportion": num_zero / len(column)}}
+    return {"zeros": {"count": int(num_zero),
+                           "proportion": float(num_zero / len(column))}}
